Add InterviewScript to every interface and simplify Interface definition
