--- src/components/StageEditor/Interfaces/AlterEdgeForm.js
@@ -1,21 +1,0 @@
-import {
-  FilteredEdgeType,
-  Form,
-  IntroductionPanel,
-} from '@components/sections';
-
-const template = {
-};
-
-const AlterEdgeForm = {
-  headerSections: [
-    FilteredEdgeType,
-  ],
-  sections: [
-    IntroductionPanel,
-    Form,
-  ],
-  template,
-};
-
-export default AlterEdgeForm;
