--- src/components/StageEditor/Interfaces/AlterForm.js
@@ -1,17 +1,0 @@
-import {
-  FilteredNodeType,
-  Form,
-  IntroductionPanel,
-} from '@components/sections';
-
-const AlterForm = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    IntroductionPanel,
-    Form,
-  ],
-};
-
-export default AlterForm;
