--- src/components/StageEditor/Interfaces/CategoricalBin.js
@@ -1,15 +1,0 @@
-import {
-  FilteredNodeType,
-  CategoricalBinPrompts,
-} from '@components/sections';
-
-const CategoricalBin = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    CategoricalBinPrompts,
-  ],
-};
-
-export default CategoricalBin;
