--- src/components/StageEditor/Interfaces/DyadCensus.js
@@ -1,17 +1,0 @@
-import {
-  FilteredNodeType,
-  IntroductionPanel,
-  DyadCensusPrompts,
-} from '@components/sections';
-
-const DyadCensus = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    IntroductionPanel,
-    DyadCensusPrompts,
-  ],
-};
-
-export default DyadCensus;
