--- src/components/StageEditor/Interfaces/EgoForm.js
@@ -1,19 +1,0 @@
-import {
-  Form,
-  IntroductionPanel,
-} from '@components/sections';
-
-const template = {
-};
-
-const EgoForm = {
-  headerSections: [
-  ],
-  sections: [
-    IntroductionPanel,
-    Form,
-  ],
-  template,
-};
-
-export default EgoForm;
