--- src/components/StageEditor/Interfaces/Information.js
@@ -1,15 +1,0 @@
-import {
-  Title,
-  ContentGrid,
-} from '@components/sections';
-
-const Information = {
-  headerSections: [
-  ],
-  sections: [
-    Title,
-    ContentGrid,
-  ],
-};
-
-export default Information;
