--- src/components/StageEditor/Interfaces/NameGenerator.js
@@ -1,22 +1,0 @@
-import {
-  NodeType,
-  Form,
-  NameGeneratorPrompts,
-  NodePanels,
-  InterviewScript,
-} from '@components/sections';
-
-const NameGenerator = {
-  headerSections: [
-    NodeType,
-  ],
-  sections: [
-    InterviewScript,
-    Form,
-    NameGeneratorPrompts,
-    NodePanels,
-  ],
-  name: 'Name Generator (using forms)',
-};
-
-export default NameGenerator;
