--- src/components/StageEditor/Interfaces/NameGeneratorAutoComplete.js
@@ -1,22 +1,0 @@
-import {
-  NodeType,
-  ExternalDataSource,
-  CardDisplayOptions,
-  SearchOptionsForExternalData,
-  NameGeneratorAutoCompletePrompts,
-} from '@components/sections';
-
-const NameGeneratorAutoComplete = {
-  headerSections: [
-    NodeType,
-  ],
-  sections: [
-    ExternalDataSource,
-    CardDisplayOptions,
-    SearchOptionsForExternalData,
-    NameGeneratorAutoCompletePrompts,
-  ],
-  name: 'Large Roster Name Generator',
-};
-
-export default NameGeneratorAutoComplete;
