--- src/components/StageEditor/Interfaces/NameGeneratorQuickAdd.js
@@ -1,20 +1,0 @@
-import {
-  NodeType,
-  QuickAdd,
-  NameGeneratorPrompts,
-  NodePanels,
-} from '@components/sections';
-
-const NameGeneratorQuickAdd = {
-  headerSections: [
-    NodeType,
-  ],
-  sections: [
-    QuickAdd,
-    NameGeneratorPrompts,
-    NodePanels,
-  ],
-  name: 'Name Generator (quick add mode)',
-};
-
-export default NameGeneratorQuickAdd;
