--- src/components/StageEditor/Interfaces/Narrative.js
@@ -1,24 +1,0 @@
-import {
-  FilteredNodeType,
-  Background,
-  NarrativePresets,
-  NarrativeBehaviours,
-} from '@components/sections';
-
-const Narrative = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    Background,
-    NarrativePresets,
-    NarrativeBehaviours,
-  ],
-  template: {
-    behaviours: {
-      allowRepositioning: true,
-    },
-  },
-};
-
-export default Narrative;
