--- src/components/StageEditor/Interfaces/OrdinalBin.js
@@ -1,15 +1,0 @@
-import {
-  FilteredNodeType,
-  OrdinalBinPrompts,
-} from '@components/sections';
-
-const OrdinalBin = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    OrdinalBinPrompts,
-  ],
-};
-
-export default OrdinalBin;
