--- src/components/StageEditor/Interfaces/Sociogram.js
@@ -1,17 +1,0 @@
-import {
-  Background,
-  FilteredNodeType,
-  SociogramPrompts,
-} from '@components/sections';
-
-const Sociogram = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    Background,
-    SociogramPrompts,
-  ],
-};
-
-export default Sociogram;
