--- src/components/StageEditor/Interfaces/TieStrengthCensus.js
@@ -1,17 +1,0 @@
-import {
-  FilteredNodeType,
-  IntroductionPanel,
-  TieStrengthCensusPrompts,
-} from '@components/sections';
-
-const TieStrengthCensus = {
-  headerSections: [
-    FilteredNodeType,
-  ],
-  sections: [
-    IntroductionPanel,
-    TieStrengthCensusPrompts,
-  ],
-};
-
-export default TieStrengthCensus;
